requirements.txt
google-cloud-texttospeech
google-cloud-speech
pydub
pygame
python-dotenv
requests
